+++ genvalidity-hspec-optics/test/Test/Validity/OpticsSpec.hs
@@ -31,10 +31,46 @@ spec = do
             ((,) <$> (negate . abs <$> genValid) <*> (negate . abs <$> genValid))
             "tuples of negative valid doubles"
             (const [])
+    describe "lensGettingProducesValidOnValid" $
+        it "holds for (_2) for doubles" $
+        lensGettingProducesValidOnValid
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+    describe "lensGettingProducesValid" $
+        it "holds for (_2) for ints" $
+        lensGettingProducesValid ((_2) :: Lens (Int, Int) (Int, Int) Int Int)
+    describe "lensGettingProducesValidOnArbitrary" $
+        it "holds for (_2) for doubles" $
+        lensGettingProducesValidOnArbitrary
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+    describe "lensGettingProducesValidOnGen" $
+        it "holds for (_2) for special generators" $
+        lensGettingProducesValidOnGen
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+            ((,) <$> (negate . abs <$> genValid) <*> (negate . abs <$> genValid))
+            (const [])
+    describe "lensSettingProducesValidOnValid" $
+        it "holds for (_2) for doubles" $
+        lensSettingProducesValidOnValid
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+    describe "lensSettingProducesValid" $
+        it "holds for (_2) for ints" $
+        lensSettingProducesValid ((_2) :: Lens (Int, Int) (Int, Int) Int Int)
+    describe "lensSettingProducesValidOnArbitrary" $
+        it "holds for (_2) for doubles" $
+        lensSettingProducesValidOnArbitrary
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+    describe "lensSettingProducesValidOnGen" $
+        it "holds for (_2) for special generators" $
+        lensSettingProducesValidOnGen
+            ((_2) :: Lens (Double, Double) (Double, Double) Double Double)
+            (abs <$> genValid)
+            (filter (0.0 >=) . shrinkValid)
+            ((,) <$> (negate . abs <$> genValid) <*> (negate . abs <$> genValid))
+            (const [])
     describe "myBoolLens" $
-        lensSpec myBoolLens -- For any unchecked value, prefer this version if you can.
+        lensSpecOnValid myBoolLens
     describe "myRationalLens" $
-        lensSpecOnValid myRationalLens -- Only for valid values
+        lensSpecOnValid myRationalLens
 
 data MyRecord = MyRecord
     { myBool :: Bool