+++ nix/release.nix
@@ -1,0 +1,2 @@
+let pkgs = import (../default.nix);
+in pkgs.validityPackages