+++ validity/CHANGELOG.md
@@ -1,2 +1,15 @@
-Unreleased changes:
-* -0 is now a valid value for Double and Float.
+# Changelog
+
+## [Unreleased]
+
+## [0.8.0.0] - 2018-08-25
+
+### Added
+* `decorateList` in `Data.Validity`
+
+### Changed
+* `-0` is now a valid value for `Double` and `Float`.
+
+## Older versions
+
+No history before version 0.8.0.0